+++ src/types/trigram-utils.d.ts
@@ -1,0 +1,3 @@
+declare module "trigram-utils" {
+  export function asTuples(value: string): Array<[string, number]>;
+}